+++ apollo-portal/src/main/java/com/ctrip/framework/apollo/portal/spi/configuration/AuthConfiguration.java
@@ -14,20 +14,16 @@
  * limitations under the License.
  *
  */
+
 package com.ctrip.framework.apollo.portal.spi.configuration;
 
 import com.ctrip.framework.apollo.common.condition.ConditionalOnMissingProfile;
 import com.ctrip.framework.apollo.core.utils.StringUtils;
-import com.ctrip.framework.apollo.portal.component.config.PortalConfig;
 import com.ctrip.framework.apollo.portal.repository.UserRepository;
 import com.ctrip.framework.apollo.portal.spi.LogoutHandler;
 import com.ctrip.framework.apollo.portal.spi.SsoHeartbeatHandler;
 import com.ctrip.framework.apollo.portal.spi.UserInfoHolder;
 import com.ctrip.framework.apollo.portal.spi.UserService;
-import com.ctrip.framework.apollo.portal.spi.ctrip.CtripLogoutHandler;
-import com.ctrip.framework.apollo.portal.spi.ctrip.CtripSsoHeartbeatHandler;
-import com.ctrip.framework.apollo.portal.spi.ctrip.CtripUserInfoHolder;
-import com.ctrip.framework.apollo.portal.spi.ctrip.CtripUserService;
 import com.ctrip.framework.apollo.portal.spi.defaultimpl.DefaultLogoutHandler;
 import com.ctrip.framework.apollo.portal.spi.defaultimpl.DefaultSsoHeartbeatHandler;
 import com.ctrip.framework.apollo.portal.spi.defaultimpl.DefaultUserInfoHolder;
@@ -44,13 +40,12 @@ import com.ctrip.framework.apollo.portal.spi.oidc.OidcUserInfoHolder;
 import com.ctrip.framework.apollo.portal.spi.springsecurity.ApolloPasswordEncoderFactory;
 import com.ctrip.framework.apollo.portal.spi.springsecurity.SpringSecurityUserInfoHolder;
 import com.ctrip.framework.apollo.portal.spi.springsecurity.SpringSecurityUserService;
-import com.google.common.collect.Maps;
+import java.util.Collections;
+import javax.sql.DataSource;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.security.oauth2.client.OAuth2ClientProperties;
 import org.springframework.boot.autoconfigure.security.oauth2.resource.OAuth2ResourceServerProperties;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.boot.web.servlet.FilterRegistrationBean;
-import org.springframework.boot.web.servlet.ServletListenerRegistrationBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
@@ -75,159 +70,12 @@ import org.springframework.security.oauth2.client.registration.InMemoryClientReg
 import org.springframework.security.provisioning.JdbcUserDetailsManager;
 import org.springframework.security.web.authentication.LoginUrlAuthenticationEntryPoint;
 
-import javax.servlet.Filter;
-import javax.sql.DataSource;
-import java.util.Collections;
-import java.util.EventListener;
-import java.util.Map;
-
 @Configuration
 public class AuthConfiguration {
 
   private static final String[] BY_PASS_URLS = {"/prometheus/**", "/metrics/**", "/openapi/**",
       "/vendor/**", "/styles/**", "/scripts/**", "/views/**", "/img/**", "/i18n/**", "/prefix-path",
       "/health"};
-
-  /**
-   * spring.profiles.active = ctrip
-   */
-  @Configuration
-  @Profile("ctrip")
-  static class CtripAuthAutoConfiguration {
-
-    private final PortalConfig portalConfig;
-
-    public CtripAuthAutoConfiguration(final PortalConfig portalConfig) {
-      this.portalConfig = portalConfig;
-    }
-
-    @Bean
-    public ServletListenerRegistrationBean redisAppSettingListner() {
-      ServletListenerRegistrationBean redisAppSettingListener = new ServletListenerRegistrationBean();
-      redisAppSettingListener
-          .setListener(listener("org.jasig.cas.client.credis.CRedisAppSettingListner"));
-      return redisAppSettingListener;
-    }
-
-    @Bean
-    public ServletListenerRegistrationBean singleSignOutHttpSessionListener() {
-      ServletListenerRegistrationBean singleSignOutHttpSessionListener = new ServletListenerRegistrationBean();
-      singleSignOutHttpSessionListener
-          .setListener(listener("org.jasig.cas.client.session.SingleSignOutHttpSessionListener"));
-      return singleSignOutHttpSessionListener;
-    }
-
-    @Bean
-    public FilterRegistrationBean casFilter() {
-      FilterRegistrationBean singleSignOutFilter = new FilterRegistrationBean();
-      singleSignOutFilter.setFilter(filter("org.jasig.cas.client.session.SingleSignOutFilter"));
-      singleSignOutFilter.addUrlPatterns("/*");
-      singleSignOutFilter.setOrder(1);
-      return singleSignOutFilter;
-    }
-
-    @Bean
-    public FilterRegistrationBean authenticationFilter() {
-      FilterRegistrationBean casFilter = new FilterRegistrationBean();
-
-      Map<String, String> filterInitParam = Maps.newHashMap();
-      filterInitParam.put("redisClusterName", "casClientPrincipal");
-      filterInitParam.put("serverName", portalConfig.portalServerName());
-      filterInitParam.put("casServerLoginUrl", portalConfig.casServerLoginUrl());
-      //we don't want to use session to store login information, since we will be deployed to a cluster, not a single instance
-      filterInitParam.put("useSession", "false");
-      filterInitParam.put("/openapi.*", "exclude");
-
-      casFilter.setInitParameters(filterInitParam);
-      casFilter
-          .setFilter(filter("com.ctrip.framework.apollo.sso.filter.ApolloAuthenticationFilter"));
-      casFilter.addUrlPatterns("/*");
-      casFilter.setOrder(2);
-
-      return casFilter;
-    }
-
-    @Bean
-    public FilterRegistrationBean casValidationFilter() {
-      FilterRegistrationBean casValidationFilter = new FilterRegistrationBean();
-      Map<String, String> filterInitParam = Maps.newHashMap();
-      filterInitParam.put("casServerUrlPrefix", portalConfig.casServerUrlPrefix());
-      filterInitParam.put("serverName", portalConfig.portalServerName());
-      filterInitParam.put("encoding", "UTF-8");
-      //we don't want to use session to store login information, since we will be deployed to a cluster, not a single instance
-      filterInitParam.put("useSession", "false");
-      filterInitParam.put("useRedis", "true");
-      filterInitParam.put("redisClusterName", "casClientPrincipal");
-
-      casValidationFilter
-          .setFilter(
-              filter("org.jasig.cas.client.validation.Cas20ProxyReceivingTicketValidationFilter"));
-      casValidationFilter.setInitParameters(filterInitParam);
-      casValidationFilter.addUrlPatterns("/*");
-      casValidationFilter.setOrder(3);
-
-      return casValidationFilter;
-    }
-
-    @Bean
-    public FilterRegistrationBean assertionHolder() {
-      FilterRegistrationBean assertionHolderFilter = new FilterRegistrationBean();
-
-      Map<String, String> filterInitParam = Maps.newHashMap();
-      filterInitParam.put("/openapi.*", "exclude");
-
-      assertionHolderFilter.setInitParameters(filterInitParam);
-
-      assertionHolderFilter.setFilter(
-          filter("com.ctrip.framework.apollo.sso.filter.ApolloAssertionThreadLocalFilter"));
-      assertionHolderFilter.addUrlPatterns("/*");
-      assertionHolderFilter.setOrder(4);
-
-      return assertionHolderFilter;
-    }
-
-    @Bean
-    public CtripUserInfoHolder ctripUserInfoHolder() {
-      return new CtripUserInfoHolder();
-    }
-
-    @Bean
-    public CtripLogoutHandler logoutHandler() {
-      return new CtripLogoutHandler();
-    }
-
-    private Filter filter(String className) {
-      Class clazz = null;
-      try {
-        clazz = Class.forName(className);
-        Object obj = clazz.newInstance();
-        return (Filter) obj;
-      } catch (Exception e) {
-        throw new RuntimeException("instance filter fail", e);
-      }
-    }
-
-    private EventListener listener(String className) {
-      Class clazz = null;
-      try {
-        clazz = Class.forName(className);
-        Object obj = clazz.newInstance();
-        return (EventListener) obj;
-      } catch (Exception e) {
-        throw new RuntimeException("instance listener fail", e);
-      }
-    }
-
-    @Bean
-    public UserService ctripUserService(PortalConfig portalConfig) {
-      return new CtripUserService(portalConfig);
-    }
-
-    @Bean
-    public SsoHeartbeatHandler ctripSsoHeartbeatHandler() {
-      return new CtripSsoHeartbeatHandler();
-    }
-  }
 
   /**
    * spring.profiles.active = auth
@@ -398,7 +246,7 @@ public class AuthConfiguration {
 
     public SpringSecurityLDAPConfigurer(final LdapProperties ldapProperties,
         final LdapContextSource ldapContextSource,
-       final LdapExtendProperties ldapExtendProperties) {
+        final LdapExtendProperties ldapExtendProperties) {
       this.ldapProperties = ldapProperties;
       this.ldapContextSource = ldapContextSource;
       this.ldapExtendProperties = ldapExtendProperties;
@@ -409,7 +257,8 @@ public class AuthConfiguration {
       if (ldapExtendProperties.getGroup() == null || StringUtils
           .isBlank(ldapExtendProperties.getGroup().getGroupSearch())) {
         FilterBasedLdapUserSearch filterBasedLdapUserSearch = new FilterBasedLdapUserSearch("",
-            ldapProperties.getSearchFilter(), ldapContextSource);
+            ldapProperties.getSearchFilter(), ldapContextSource
+        );
         filterBasedLdapUserSearch.setSearchSubtree(true);
         return filterBasedLdapUserSearch;
       }
@@ -418,7 +267,8 @@ public class AuthConfiguration {
           ldapProperties.getBase(), ldapProperties.getSearchFilter(), ldapExtendProperties.getGroup().getGroupBase(),
           ldapContextSource, ldapExtendProperties.getGroup().getGroupSearch(),
           ldapExtendProperties.getMapping().getRdnKey(),
-          ldapExtendProperties.getGroup().getGroupMembership(),ldapExtendProperties.getMapping().getLoginId());
+          ldapExtendProperties.getGroup().getGroupMembership(), ldapExtendProperties.getMapping().getLoginId()
+      );
       filterLdapByGroupUserSearch.setSearchSubtree(true);
       return filterLdapByGroupUserSearch;
     }
@@ -445,9 +295,9 @@ public class AuthConfiguration {
           .antMatchers(BY_PASS_URLS).permitAll()
           .antMatchers("/**").authenticated();
       http.formLogin().loginPage("/signin").defaultSuccessUrl("/", true).permitAll().failureUrl("/signin?#/error").and()
-              .httpBasic();
+          .httpBasic();
       http.logout().logoutUrl("/user/logout").invalidateHttpSession(true).clearAuthentication(true)
-              .logoutSuccessUrl("/signin?#/logout");
+          .logoutSuccessUrl("/signin?#/logout");
       http.exceptionHandling().authenticationEntryPoint(new LoginUrlAuthenticationEntryPoint("/signin"));
     }