+++ CHANGES.md
@@ -9,7 +9,7 @@ Apollo 1.10.0
 * [Fix issue that the $ symbol is not used when reading shell variables](https://github.com/ctripcorp/apollo/pull/3890)
 * [Bump xstream from 1.4.17 to 1.4.18](https://github.com/apolloconfig/apollo/pull/3916)
 * [switch apollo.config-service log from warning to info level](https://github.com/ctripcorp/apollo/pull/3884)
+* [remove ctrip profile](https://github.com/ctripcorp/apollo/pull/3920)
 
 ------------------
 All issues and pull requests are [here](https://github.com/ctripcorp/apollo/milestone/8?closed=1)
-